Rewrite TextField and ReadonlyField
