+++ packages/bento-design-system/src/ReadOnlyField/ReadOnlyField.tsx
@@ -1,0 +1,16 @@
+import { TextField, TextFieldProps } from "../TextField/TextField";
+
+type Props = Omit<
+  TextFieldProps,
+  "onChange" | "onBlur" | "disabled" | "isReadOnly" | "placeholder" | "issues"
+>;
+
+const constVoid = () => {};
+
+/**
+ * A convenient wrapper to the TextField, that forces isReadOnly=true
+ * and doesn't require all those props that don't make sense for a read-only field (onChange, onBlur, placeholder, ...)
+ */
+export function ReadOnlyField(props: Props) {
+  return <TextField {...props} onChange={constVoid} onBlur={constVoid} placeholder="" isReadOnly />;
+}
